MP1-4863: Add New Collection from MP GUI MP1-4863: Remove Selected Collection from MP GUI MP1-4863: When delete Movie from User Group/Collection, if more then one User Group/Collection show Dialog MP1-4863: Add to TMDB_DE grabber Awards and MPAA Text support (1.17 and >) MP1-4863: New Language id 1337 and 1338 in EN,RU

+++ mediaportal/WindowPlugins/GUIVideos/GUIVideoTitle.cs
@@ -1,6 +1,6 @@
-#region Copyright (C) 2005-2011 Team MediaPortal
+#region Copyright (C) 2005-2017 Team MediaPortal
 
-// Copyright (C) 2005-2011 Team MediaPortal
+// Copyright (C) 2005-2017 Team MediaPortal
 // http://www.team-mediaportal.com
 // 
 // MediaPortal is free software: you can redistribute it and/or modify
@@ -636,7 +636,11 @@ namespace MediaPortal.GUI.Video
         case 1334: // Remove from Collection
           OnRemoveFromCollection(movie, itemNo);
           break;
-        
+
+        case 1337: // Add New Collection
+          OnAddCollection();
+          break;
+
         case 1270: // Add to user group
           OnAddToUserGroup(movie, itemNo);
           break;
@@ -1286,10 +1290,15 @@ namespace MediaPortal.GUI.Video
       {
         dlg.AddLocalizedString(1293); //Search movie
 
+        if (handler.CurrentLevelWhere == "movie collections" || handler.CurrentLevelWhere == "movie collections only")
+        {
+          dlg.AddLocalizedString(1337); // Add Collection
+          dlg.AddLocalizedString(1338); // Remove selected Collection
+        }
         if (handler.CurrentLevelWhere == "user groups" || handler.CurrentLevelWhere == "user groups only")
         {
-          dlg.AddLocalizedString(1272); //Add usergroup
-          dlg.AddLocalizedString(1273); //Remove selected usergroup
+          dlg.AddLocalizedString(1272); // Add usergroup
+          dlg.AddLocalizedString(1273); // Remove selected usergroup
         }
       }
 
@@ -1342,7 +1351,19 @@ namespace MediaPortal.GUI.Video
 
           OnRemoveUserGroup(item.Label);
           break;
-        
+        case 1337: // Add New Collection
+          OnAddCollection();
+          break;
+        case 1338: // Remove Collection
+          GUIListItem selitem = facadeLayout.SelectedListItem;
+
+          if (selitem == null)
+          {
+            return;
+          }
+
+          OnRemoveCollection(selitem.Label);
+          break;
       }
     }
 
@@ -2270,8 +2291,49 @@ namespace MediaPortal.GUI.Video
 
     private void OnRemoveFromCollection(IMDBMovie movie, int itemIndex)
     {
-      // string group = m_history.Get("movie collections");
-      string strCollection = m_history.Get(handler.CurrentLevelWhere);
+      ArrayList movieCollections = new ArrayList();
+      VideoDatabase.GetMovieCollections(movie.ID, movieCollections);
+      if (movieCollections.Count == 0)
+      {
+        return;
+      }
+
+      string strCollection = string.Empty;
+      if (movieCollections.Count > 1)
+      {
+        GUIDialogMenu dlg = (GUIDialogMenu)GUIWindowManager.GetWindow((int)Window.WINDOW_DIALOG_MENU);
+
+        if (dlg == null)
+        {
+          return;
+        }
+
+        dlg.Reset();
+        dlg.SetHeading(1334); // Remove from Collection
+
+        foreach (string strColl in movieCollections)
+        {
+          dlg.Add(strColl);
+        }
+
+        dlg.DoModal(GetID);
+
+        if (dlg.SelectedId == -1)
+        {
+          return;
+        }
+        strCollection = dlg.SelectedLabelText;
+      }
+      else
+      {
+        strCollection = movieCollections[0].ToString();
+      }
+
+      if (string.IsNullOrEmpty(strCollection))
+      {
+        return;
+      }
+
       VideoDatabase.RemoveCollectionFromMovie(movie.ID, VideoDatabase.AddCollection(strCollection));
 
       currentSelectedItem = itemIndex;
@@ -2291,6 +2353,26 @@ namespace MediaPortal.GUI.Video
           GUIControl.SelectItemControl(GetID, facadeLayout.GetID, currentSelectedItem);
         }
       }
+    }
+
+    private void OnAddCollection()
+    {
+      string newCollection = string.Empty;
+      VirtualKeyboard.GetKeyboard(ref newCollection, GetID);
+
+      if (string.IsNullOrEmpty(newCollection))
+      {
+        return;
+      }
+
+      VideoDatabase.AddCollection(newCollection);
+      LoadDirectory(currentFolder);
+    }
+
+    private void OnRemoveCollection(string collection)
+    {
+      VideoDatabase.DeleteCollection(collection);
+      LoadDirectory(currentFolder);
     }
 
     private void OnAddToUserGroup(IMDBMovie movie, int itemIndex)
@@ -2348,8 +2430,49 @@ namespace MediaPortal.GUI.Video
 
     private void OnRemoveFromUserGroup(IMDBMovie movie, int itemIndex)
     {
-      // string group = m_history.Get("user groups");
-      string group = m_history.Get(handler.CurrentLevelWhere);
+      ArrayList movieGroups = new ArrayList();
+      VideoDatabase.GetMovieUserGroups(movie.ID, movieGroups);
+      if (movieGroups.Count == 0)
+      {
+        return;
+      }
+
+      string group = string.Empty;
+      if (movieGroups.Count > 1)
+      {
+        GUIDialogMenu dlg = (GUIDialogMenu)GUIWindowManager.GetWindow((int)Window.WINDOW_DIALOG_MENU);
+
+        if (dlg == null)
+        {
+          return;
+        }
+
+        dlg.Reset();
+        dlg.SetHeading(1271); // Remove from user group
+
+        foreach (string strGroup in movieGroups)
+        {
+          dlg.Add(strGroup);
+        }
+
+        dlg.DoModal(GetID);
+
+        if (dlg.SelectedId == -1)
+        {
+          return;
+        }
+        group = dlg.SelectedLabelText;
+      }
+      else
+      {
+        group = movieGroups[0].ToString();
+      }
+
+      if (string.IsNullOrEmpty(group))
+      {
+        return;
+      }
+
       VideoDatabase.RemoveUserGroupFromMovie(movie.ID, VideoDatabase.AddUserGroup(group));
       
       currentSelectedItem = itemIndex;
